split module up
